修改 修改项目进度显示 优化 优化项目详细页面 文件显示

+++ src/main/java/com/xh/service/impl/FileServiceImpl.java
@@ -339,7 +339,7 @@ public class FileServiceImpl extends BaseService implements IFileService {
 						}
 						ratio = completedProject + "/" + (progressProject + completedProject);
 						per = (progressProject + completedProject) > 0
-								? (int) (completedProject / (progressProject + completedProject) * 100)
+								? (int) ((float) completedProject / (float) (progressProject + completedProject) * 100)
 								: 0;
 					}
 					// 3.该员工所具有权限的文件信息
